modif: eliminado menu_poketest y pasado al main

+++ src/juego_poketest.h
@@ -10,34 +10,34 @@
 #include <stdio.h>
 #include <unistd.h>
 
-enum color_tarjeta {COLOR_AZUL, COLOR_ROJO};
+enum color_tarjeta { COLOR_AZUL, COLOR_ROJO };
 
 typedef struct tarjeta {
-    int numero_tarjeta;
-    int id_pokemon;
-    enum color_tarjeta color_actual;
-    struct pokemon *valor;
-} tarjeta_t;  
+	size_t numero_tarjeta;
+	int id_pokemon;
+	enum color_tarjeta color_actual;
+	struct pokemon *valor;
+} tarjeta_t;
 
 typedef struct registro_historial {
-    char *registro;
-    int primer_carta, segunda_carta;
-    bool resultado;
+	char *registro;
+	int primer_carta, segunda_carta;
+	bool resultado;
 } registro_historial_t;
 
 typedef struct jugador {
-    size_t puntaje;
+	size_t puntaje;
 } jugador_t;
 
 typedef struct juego_poketest {
-    tp1_t *archivo_pokemones;
-    lista_t *historial;
-    tarjeta_t *tarjetas;
-    size_t semilla;
-    size_t cantidad_cargada;
-    size_t cantidad_tarjetas;
-    jugador_t jugador_1, jugador_2;
-    bool turno_jugador_1;
+	tp1_t *archivo_pokemones;
+	lista_t *historial;
+	tarjeta_t *tarjetas;
+	int semilla;
+	size_t cantidad_cargada;
+	size_t cantidad_tarjetas;
+	jugador_t jugador_1, jugador_2;
+	bool turno_jugador_1;
 } juego_poketest_t;
 
 /*
@@ -45,7 +45,7 @@ typedef struct juego_poketest {
 *   Si la semilla dada es menor a 9 lo crea con ese valor como semilla.
 *   Devuelve NULL en caso de error o la estructura en caso de exito.
 */
-juego_poketest_t *juego_poketest_crear(tp1_t *archivo_pokemones, size_t semilla);
+juego_poketest_t *juego_poketest_crear(tp1_t *archivo_pokemones, int semilla);
 
 /*
 *   Carga cierto contenido del archivo de pokemones para usar en el juego.
